Fix crash after running over first dataset
Seems to be caused by deleting the cloneTrees. This makes sense
as they weren't created using new but it is a little bizarre
that changing where outfile was declared induced this behaviour...

While I was working on this I wrapped outFile in a smart pointer,
which may as well stay in.